added general MERS and refactored MERS folder
